requirements.txt
Django==1.5.1
Fabric==1.6.1
argparse==1.2.1
djangorestframework==2.3.6
gunicorn==17.5
paramiko==1.10.1
psycopg2==2.5.1
pycrypto==2.6
wsgiref==0.1.2
